record_recogniser/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Django==5.2.7
filelock==3.20.0
fonttools==4.60.1
fsspec==2025.10.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==12.0.0
polars==1.35.1
polars-runtime-32==1.35.1
psutil==7.1.3
pyparsing==3.2.5
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
torch==2.9.0
torchaudio==2.9.0
torchvision==0.24.0
typing_extensions==4.15.0
ultralytics==8.3.223
ultralytics-thop==2.0.18
urllib3==2.5.0
